add static_cast to fix compiler warning

+++ rocprim/include/rocprim/device/device_radix_sort.hpp
@@ -722,7 +722,8 @@ hipError_t
     }
     // For sizeof(key_type) <= 2, onesweep is 2x/3x faster (also with values) when
     // input_size > 100K, so don't use radix_sort_merge_sort then.
-    else if(size <= merge_sort_limit && (sizeof(key_type) > 2 || size < 100000))
+    else if(static_cast<size_t>(size) <= merge_sort_limit
+            && (sizeof(key_type) > 2 || size < 100000))
     {
         is_result_in_output = true;
         // note: Config::merge_sort_config may be default_config
